refactor: update project structure and configuration; remove unused files, adjust paths, and enhance build scripts

+++ tsup.types.ts
@@ -6,7 +6,7 @@ import { defineConfig } from 'tsup'
 // 排除掉`package.json`文件不删除 输出目录其他文件删除
 const files = fs.readdirSync(new URL('packages/types', import.meta.url))
 files.forEach((file) => {
-  if (file !== 'package.json') {
+  if (file !== 'package.json' && file !== 'index.d.ts') {
     fs.rmSync(new URL(`packages/types/${file}`, import.meta.url), { recursive: true })
   }
 })
